requirements.txt
flax==0.10.6
gymnasium==1.1.1
jax==0.6.0
matplotlib==3.10.3
neat-python @ git+https://github.com/CodeReclaimers/neat-python.git@37bc8bb73fd6153a115001c2646f9f02bac3ad81
networkx==3.4.2
numpy==2.2.5
optax==0.2.4
pygame==2.6.1
stable_baselines3==2.6.0
tensorflow_probability==0.25.0
wandb==0.19.11
forward @ git+https://github.com/BoCtrl-C/forward.git@16a8dc4421f62603321ffc134a31c7f82ec711b8
sbx-rl @ git+https://github.com/araffin/sbx@1c79684e45a5692063e28f3bdecf9fd65dd8d24d
